+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -2,6 +2,7 @@
 --------------------
 * portal_map methods are no longer called if the content type is not
   geo_enabled [dumitval]
+* gl_changeLanguage properly redirects when no referer [simiamih]
 
 2.12.30 (2011-11-16)
 --------------------